added message class for pretty printing
