+++ include/cutee/test_failed.hpp
@@ -7,85 +7,12 @@
 #include<type_traits>
 #include<iomanip>
 
+#include "message.hpp"
 #include "failed_data.hpp"
 #include "float_eq.hpp"
 
 namespace cutee
 {
-namespace detail
-{
-
-template<class T>
-struct is_complex
-   :  public std::false_type
-{
-};
-
-template<class T>
-struct is_complex<std::complex<T>>
-   :  public std::true_type
-{
-   static_assert(std::is_floating_point<T>::value, "T must be floating-point type.");
-};
-
-
-template<bool>
-struct output_float_distance
-   :  public std::false_type
-{
-   template<class T>
-   static void apply(std::stringstream& s, const T& got, const T& expected)
-   {
-      // do nothing
-   }
-
-   template<class T, class U>
-   static void apply(std::stringstream& s, const T& got, const U& expected)
-   {
-      // do nothing
-   }
-};
-
-// specialization for floating point
-template<>
-struct output_float_distance<true>
-   :  public std::true_type
-{
-   template
-      <  class T
-      ,  std::enable_if_t<std::is_floating_point<T>::value, void*> = nullptr
-      >
-   static void apply(std::stringstream& s, const T& got, const T& expected)
-   {
-      s << " dist: " << numeric::float_ulps(expected, got);
-   }
-   
-   template
-      <  class T
-      ,  std::enable_if_t<std::is_floating_point<T>::value, void*> = nullptr
-      >
-   static void apply(std::stringstream& s, const std::complex<T>& got, const std::complex<T>& expected)
-   {
-      s  << " dist: (" 
-         << numeric::float_ulps(expected.real(), got.real()) 
-         << ", "
-         << numeric::float_ulps(expected.imag(), got.imag()) 
-         << ")"
-         ;
-   }
-
-   template
-      <  class T
-      ,  class U
-      >
-   static void apply(std::stringstream& s, const T& got, const U& expected)
-   {
-      //static_assert(std::is_same<T, U>::value, "not the same types.");
-      s << " dist : N/A";
-   }
-};
-
-} /* namespace detail */
                               
 
 //
